Changes the way validation is done.
Now, the user's |url| parameters are completely ignored. The original
url fields are all that matters. We go through Google image search
results, looking for one that matches the original url the user sent
us, and then we download the thumbnail URL from that result. This
fixes a security hole my spot fix a few changes ago opened up.

Theoretically, this makes deferring the porn checking harder, as we
actually use information from the search results to construct the
entity, not just validate it. However, I'm quite sure we can speed up
the search by making it cut out once it finds a match, and if it's
still bad after that, we can hack some other way together. For now,
this is better as it doesn't rely on any assumptions about the Google
Image Search thumbnail URLs.