Gruntfile isn't working for bootstrap.css.
I don't care to debug this right now, so I added shell:copy_css to work
around it.

We are almost certainly missing angular.js from the build/ output
directory at this point. We could replace it with another 'shell'
section if we have to. But we only use it in /make anyway, IIRC.